adicionando o model usuário no repository e criando o usuário
